flashy/requirements.txt
ray[tune]
lightning-flash[image,text]
fiftyone
click=8
streamlit
jinja2
